producer/requirements.txt
kafka-python
pymongo
python-dotenv
